feat: rearrange and break up
